requirements.txt
aiogram
telethon
google-genai
python-dotenv